# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ero Server 's hmp BackEnd",
        "title": "Ero BackEnd API",
        "termsOfService": "https://api.ero.ink",
        "contact": {
            "name": "API Support",
            "url": "https://9bie.org",
            "email": "blackguwc@163.com"
        },
        "license": {
            "name": "What the Fuck PL",
            "url": "https://ero.ink"
        },
        "version": "1.0"
    },
    "host": "api.ero.ink",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/archive/": {
            "get": {
                "description": "获取单个文章的详细信息",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "archive"
                ],
                "summary": "获取文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.ArchiveResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "必须要登陆且创建者为自己或者自己权限是管理员",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "archive"
                ],
                "summary": "更新文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日文标题",
                        "name": "japanese_title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "作者",
                        "name": "author",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "文章内容",
                        "name": "content",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "隐藏内容",
                        "name": "primary_content",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "封面，这个是个url地址",
                        "name": "cover",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "标签，推荐使用/分割，例如 纯爱/治愈/等等 ",
                        "name": "tag",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.ArchiveResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "必须要登陆",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "archive"
                ],
                "summary": "创建文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日文标题",
                        "name": "japanese_title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "作者",
                        "name": "author",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "文章内容",
                        "name": "content",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "隐藏内容",
                        "name": "primary_content",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "封面，这个是个url地址",
                        "name": "cover",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "标签，推荐使用/分割，例如 纯爱/治愈/等等 ",
                        "name": "tag",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.ArchiveResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "必须要登陆且创建者为自己或者管理员",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "archive"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/download/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "必须要登陆，文件ID就是在这里去拿的，图片等静态 文件可以直接访问/img/*获得，novel必须要经过/download",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file",
                    "volume"
                ],
                "summary": "上传小说分卷",
                "parameters": [
                    {
                        "type": "string",
                        "description": "从/api/v1/novel/volume/:id得来",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "从/api/v1/novel/volume/:id得来",
                        "name": "filename",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "从/api/v1/novel/volume/:id得来",
                        "name": "hash",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.FileResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/novel/": {
            "get": {
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novel"
                ],
                "summary": "获取小说详细",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.NovelResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novel"
                ],
                "summary": "更新小说",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说ID",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "小说标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "小说封面",
                        "name": "cover",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否完结",
                        "name": "Ended",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "目标等级：还未实现",
                        "name": "level",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.NovelResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "接收者必须为自己或者用户权限为管理员",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novel"
                ],
                "summary": "删除小说",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "serializer.Archive": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "create_id": {
                    "type": "integer"
                },
                "create_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "japanese_title": {
                    "type": "string"
                },
                "primary_content": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "serializer.ArchiveResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/serializer.Archive"
                },
                "error": {
                    "type": "string",
                    "example": "Error."
                },
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "serializer.File": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "serializer.FileResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/serializer.File"
                },
                "error": {
                    "type": "string",
                    "example": "Error."
                },
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "serializer.Novel": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "ended": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "subscribed": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "serializer.NovelResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/serializer.Novel"
                },
                "error": {
                    "type": "string",
                    "example": "Error."
                },
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "serializer.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string",
                    "example": "Error."
                },
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
    }
}